+++ api/models/Recruiter.js
@@ -1,46 +1,32 @@
-const S = require("sequelize");
-const db = require("../config/db");
-
-class Recruiter extends S.Model {}
-
-Recruiter.init(
-  {
-    name: {
-      type: S.STRING,
-      allowNull: false,
-    },
-    email: {
-      type: S.STRING,
-      allowNull: false,
-      validate: {
-        isEmail: true,
-      },
-    },
-    lastName: {
-      type: S.STRING,
-      allowNull: false,
-    },
-    country: {
-      type: S.STRING,
-      allowNull: true,
-    },
-    description: {
-      type: S.TEXT,
-      allowNull: true,
-    },
-    experienceField: {
-      type: S.STRING,
-      allowNull: true,
-    },
-    rating: {
-      type: S.INTEGER,
-      allowNull: true,
-    },
-  },
-  {
-    sequelize: db,
-    modelName: "recruiter",
+"use strict";
+const { Model } = require("sequelize");
+module.exports = (sequelize, DataTypes) => {
+  class Recruiter extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      // define association here
+    }
   }
-);
-
-module.exports = Recruiter;
+  Recruiter.init(
+    {
+      email: DataTypes.STRING,
+      password: DataTypes.STRING,
+      name: DataTypes.STRING,
+      lastName: DataTypes.STRING,
+      country: DataTypes.STRING,
+      description: DataTypes.TEXT,
+      experienceField: DataTypes.STRING,
+      rating: DataTypes.INTEGER,
+      activeSearchs: DataTypes.INTEGER,
+    },
+    {
+      sequelize,
+      modelName: "Recruiter",
+    }
+  );
+  return Recruiter;
+};
