--- api/models/User.js
@@ -1,33 +1,0 @@
-const S = require("sequelize");
-const db = require("../config/db");
-
-class User extends S.Model {}
-
-User.init(
-  {
-    email: {
-      type: S.STRING,
-      allowNull: false,
-      validate: {
-        isEmail: true,
-      },
-    },
-    role: {
-      type: S.STRING,
-      defaultValue: "admin",
-    },
-    password: {
-      type: S.STRING,
-      allowNull: false,
-    },
-    salt: {
-      type: S.STRING,
-    },
-  },
-  {
-    sequelize: db,
-    modelName: "user",
-  }
-);
-
-module.exports = User;
